Cleanup cache cleaner pruneCollection function - use size_t instead of unsigned int for sizes - const correctness - layout - simplify seond loop and use return value of erase(), it points to the next value

+++ pdns/recpacketcache.cc
@@ -212,7 +212,7 @@ uint64_t RecursorPacketCache::bytes()
   return sum;
 }
 
-void RecursorPacketCache::doPruneTo(unsigned int maxCached)
+void RecursorPacketCache::doPruneTo(size_t maxCached)
 {
   pruneCollection<SequencedTag>(*this, d_packetCache, maxCached);
 }
